+++ src/flask_exts/template/base.py
@@ -1,12 +1,9 @@
 import os.path as op
 from flask import Blueprint
+from .dataset import Dataset
+from .funcs import Funcs
 from .theme import BootstrapTheme
-from .utils import type_name
-from .utils import is_hidden_field
-from .utils import is_required_form_field
-from .utils import get_table_titles
-from .form.csrf import get_or_generate_csrf_token as csrf_token
-
+from .plugin import Plugin
 
 class Template:
     """Template extension for Flask applications."""
@@ -19,16 +16,12 @@ class Template:
     def init_app(self, app):
         self.app = app
         self.init_blueprint(app)
-
+        app.jinja_env.globals["template"] = self
+        self.dataset = Dataset()
+        self.funcs = Funcs()
         self.theme = self.get_theme()
-        app.jinja_env.globals["template"] = self.theme
-
-        app.jinja_env.globals["type_name"] = type_name
-        app.jinja_env.globals["is_hidden_field"] = is_hidden_field
-        app.jinja_env.globals["is_required_form_field"] = is_required_form_field
-        app.jinja_env.globals["get_table_titles"] = get_table_titles
-        app.jinja_env.globals["csrf_token"] = csrf_token
-
+        self.plugin = Plugin()
+        
     def get_theme(self):
         return BootstrapTheme()
 
@@ -42,7 +35,3 @@ class Template:
             # static_url_path='/template/static',
         )
         app.register_blueprint(blueprint)
-
-        # @app.context_processor
-        # def get_template():
-        #     return {"template": theme}