+++ src/flask_exts/template/forms/profile.py
@@ -1,6 +1,6 @@
 from wtforms import StringField
 from wtforms import SubmitField
-from ..template.form.flask_form import FlaskForm
+from ..form.flask_form import FlaskForm
 
 
 class ProfileForm(FlaskForm):