+++ src/flask_exts/template/forms/two_factor.py
@@ -2,7 +2,7 @@ from wtforms import StringField
 from wtforms import SubmitField
 from wtforms.validators import DataRequired
 from wtforms.validators import Length
-from ..template.form.flask_form import FlaskForm
+from ..form.flask_form import FlaskForm
 
 
 class TwoFactorForm(FlaskForm):